implement server-side WRQ; improve internals
This commit adds support for handling WRQ requests from clients to store
a file on the TFTP server. Similar to the RRQ implementation, this was
optimized to have a low footprint and to be quite fast.

Adding support also required several changes to the internals, which
also have been improved a lot. The process of receiving files and
sending files is each encapsulated to be usable for the client
implementation in the future (since the file transfer process doesn't
differ).

Signed-off-by: Jonas Jelonek <jonas.jelonek@protonmail.ch>