bump to 1.0.0
This version bump marks the stable state of the project. During the last
commits, support for most parts of the TFTP protocol was added and is
working so far without any major issues.
There are still some things to do, however, they are rather minor things
not affected the functionality or the external usage of the application.

Signed-off-by: Jonas Jelonek <jonas.jelonek@protonmail.ch>