allow use of stack-allocated buffer with packet builder
This changes the packet builders to optionally be used with an existing
buffer that may be stack-allocated. A reference to such a buffer can be
added to the builder and will be used when calling build() instead of
the heap-allocated Vec<> that is created by default.

The user of this feature needs to make sure that the buffer is large
enough. The build() method will silently fail in case the buffer is not
large enough, maybe causing a corrupted packet!

Signed-off-by: Jonas Jelonek <jonas.jelonek@protonmail.ch>