remove usage of unstable iter_advance_by feature
This removes the usage of the iter_advance_by feature that hasn't been
stabilized yet in Rust. This was used in the TftpReq struct to skip two
elements of an iterator used for parsing the options and was one reason
for the need of nightly toolchain.
Although we still need to advance the iterator, we can use the already
stabilized nth() feature which works similar but might be less intuitive.

Signed-off-by: Jonas Jelonek <jonas.jelonek@protonmail.ch>